add pagination in user tables

+++ admin/src/components/User_data/Contact_User.jsx
@@ -42,6 +42,13 @@ function  Contact_User() {
 
 
     const handleDelete = async(id)=>{
+        // Display a confirmation popup
+    const isConfirmed = window.confirm('Are you sure you want to Delete?');
+
+    if (!isConfirmed) {
+      // If the user cancels the deletion, do nothing
+      return;
+    }
       try{
         const res = await axios.delete(`https://bharatroofers.com/api/property/deleteContactedUser/${id}`,
         {
@@ -65,6 +72,27 @@ function  Contact_User() {
         console.log(err)
       }
     }
+
+    const [currentPage, setCurrentPage] =
+    useState(1);
+  const studentsPerPage = 10;
+
+    const nextPage = () =>
+    setCurrentPage(currentPage + 1);
+  const prevPage = () =>
+    setCurrentPage(currentPage - 1);
+
+
+    const indexOfLastStudent =
+    currentPage * studentsPerPage;
+  const indexOfFirstStudent =
+    indexOfLastStudent -
+    studentsPerPage;
+  const currentuser =
+    user.slice(
+      indexOfFirstStudent,
+      indexOfLastStudent
+    );
  
   return (
    <Wrapper>
@@ -82,11 +110,12 @@ function  Contact_User() {
    <div className="col-lg-12">
   <div className="widget-area-2 proclinic-box-shadow " id='tableres'>
                     <h5 className="widget-title" id='title'>Contacted User</h5>
+                    <h5 className="total" >Total - {user?.length}</h5>
                     <div className="table-responsive">
                         <table className="table table-bordered table-striped">
                           <thead>
                             <tr>
-                              <th>User id</th>
+                              <th>No.</th>
                               <th>User Name</th>
                               <th>Phone Number</th>
                               <th>Email</th>
@@ -102,11 +131,12 @@ function  Contact_User() {
                           
                       {user && user?.length > 0 ?  
                       <>
-                     {user && user?.length > 0 &&  user?.map((user) => {
+                     {user && user?.length > 0 &&  currentuser?.map((user,index) => {
+                      const serialNumber = indexOfFirstStudent + index + 1; // Calculate serial number
                       return (
                         <tbody key={user?.id}>
                         <tr>
-                          <td>{user?.id}</td>
+                          <td>{serialNumber}</td>
                           <td>{user?.name}</td>
                           <td>{user?.phone}</td>
                           <td>{user?.email}</td>
@@ -136,6 +166,30 @@ function  Contact_User() {
                      }
                      </table>
                       </div>
+                      <div className="pagination-section mt-3">
+              <div className="pagination">
+                <button
+                  onClick={prevPage}
+                  disabled={
+                    currentPage === 1
+                  }
+                  className="btn btn-danger">
+                  Previous
+                </button>
+                <span className="fs-4 mx-3">
+                  {currentPage}
+                </span>
+                <button
+                  onClick={nextPage}
+                  className="btn btn-success"
+                  disabled={
+                    currentuser.length <
+                    studentsPerPage
+                  }>
+                  Next
+                </button>
+              </div>
+            </div>
                   </div>
    </div>
   
@@ -224,6 +278,12 @@ const Wrapper = styled.div`
   
   margin-left: 2rem;
   }
+  }
+  .total{
+    @media screen and (max-width: 768px) {
+    margin-left: 2rem;
+  }
+    
   }
 
 `
